WEP-47 refactor WeeksPage content components
